requirements.txt
Flask==3.0.3
python-pptx==0.6.23
requests==2.32.3
